refactor: PR #169 cleanup - simplify status machine + improve UX
## Status Machine Simplification
- Remove 'processing' status from order flow
- Streamline transitions: draft→pending→paid→shipped→delivered
- Cancel only allowed from pending/paid states
- Update validation logic and UI components

## Test Structure Improvement
- Split large order-management.spec.ts (324 LOC) into focused files:
  - order-admin.spec.ts (admin workflow tests)
  - order-producer.spec.ts (producer workflow tests)
  - order-validation.spec.ts (status validation tests)
- Use data-testid selectors for deterministic testing
- Remove 'processing' references from mock data

## UX Enhancement
- Replace alert() with toast notifications using existing ToastContext
- Success messages for order status updates
- Error messages with Greek localization
- Immediate visual feedback for user actions

## Documentation Organization
- Move reports to docs/reports/2025-09-16/ structure
- Maintain proper documentation hierarchy
